requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
eventlet==0.34.2
exceptiongroup==1.2.0
greenlet==3.0.3
iniconfig==2.0.0
kombu==5.3.4
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==1.7.1
pytest==7.4.3
pytest-celery==0.0.0
pytest-django==4.7.0
pytest-mock==3.12.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
